+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - unsupported device language fallback properly to english
+- podcast episode sessions resume from last progress
 
 ## [v0.6.2] - 2026-08-23